# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hemera API
  version: 1.0.0
  description: |
    OpenAPI 3.1 specification for all Hemera platform endpoints.
    Covers public, authenticated, admin, webhook, and monitoring routes.
servers:
  - url: http://localhost:3000/api
    description: Local Development
  - url: https://staging.hemera.app/api
    description: Staging Environment
  - url: https://hemera.app/api
    description: Production
security:
  - clerkAuth: []
...
paths:
...
  /users/{id}:
    get:
      tags: [Admin]
      operationId: getUser
      summary: Get user details
      description: Returns user profile details for admin tooling.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: cuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
      security:
        - clerkAuth: []
    put:
      tags: [Admin]
      operationId: updateUser
      summary: Update user
      description: Updates user profile attributes or role.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: cuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
                role:
                  type: string
                  enum: [USER, ADMIN]
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
      security:
        - clerkAuth: []
    delete:
      tags: [Admin]
      operationId: deleteUser
      summary: Delete user
      description: Removes the specified user and related role assignments.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: cuid
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
      security:
        - clerkAuth: []
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Clerk JWT token issued after authentication.
  schemas:
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data: {}
        requestId:
          type: string
          format: uuid
      required: [success, requestId]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: cuid
        clerkId:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clerkId, email, role]
...